requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
pytest==7.4.4
pytest-mock==3.12.0
